Sprint 2 Phase 1: delete Sprint 1.5 info graphic UI
DELETE Sprint 1.5 v4 UI (being replaced by Napkin document viewer in Sprint 2):
- src/present/deck-view.js (info graphic deck view)
- src/present/info-graphic-render.js (silhouette + chrome compositor)
- src/present/linear-layout.js (deck-level region layout)
- scripts/test-info-graphic-render.mjs
- scripts/test-linear-layout.mjs

Remove the 2 test entries from scripts/run-tests.mjs.

pipeline.js still consumed computeRegions from linear-layout.js (asserted by
test-pipeline.mjs: every variant has region). Inlined the minimal
computeBoundingBox + computeRegions helpers directly into pipeline.js so the
pipeline stays self-contained. Region remains attached to each variant for
backwards-compat with downstream SceneData consumers — Sprint 2 document
viewer does not use it for layout.

After Phase 1: HTML router still references deleted deck-view.js (broken).
Document viewer will be wired in Phase 6/7. Branch is intentionally
non-loadable in browser between Phase 1 and Phase 7.

npm test 32/32 (down from 34; 2 deleted test files removed).
